refactor: change structure for cart
