+++ src/pages/cart/index.html
@@ -1,5 +1,4 @@
 <section class="cart__wrapper">
-
     <div class="cart">
         <div class="cart__list-wrapper">
             <h2 class="cart__title">Корзина</h2>
@@ -20,9 +19,9 @@
 					<div class="promokod">
 						<p class="promokod__text">Введите промокод:</p>
 						<div class="promokod__search">
-							
+							<!--
 							<input class="search-input promokod-search" type="search">
-							
+							-->
 						</div>
 						<div class="promokod__offer">
 							<!--
@@ -42,8 +41,6 @@
             </div>
         </div>
     </div>
-
-
 </section>
